% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_border.R
\name{add_border}
\alias{add_border}
\title{(Internal) Set a Border}
\usage{
add_border(obj1, obj2)
}
\arguments{
\item{obj1}{The existing \code{ssbrain} object}

\item{obj2}{The new \code{ssborder} object to add}
}
\description{
The internal function that calculates the result for \code{\link[ssbrain]{ss_border}}
}
