% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_seed.R
\name{add_dconn}
\alias{add_dconn}
\title{(Internal) Set a Seed}
\usage{
add_dconn(obj1, obj2)
}
\arguments{
\item{obj1}{The existing \code{ssbrain} object}

\item{obj2}{The new \code{ssdconn} object to add}
}
\description{
The internal function that calculates the result for \code{\link[ssbrain]{ss_seed}}
}
