% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_dlabel.R
\name{add_dlabel}
\alias{add_dlabel}
\title{(Internal) Set a Dlabel}
\usage{
add_dlabel(obj1, obj2)
}
\arguments{
\item{obj1}{The existing \code{ssbrain} object}

\item{obj2}{The new \code{ssdlabel} object to add}
}
\description{
The internal function that calculates the result for \code{\link[ssbrain]{ss_dlabel}}
}
