% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_dscalar.R
\name{add_dscalar}
\alias{add_dscalar}
\title{(Internal) Set a Dscalar}
\usage{
add_dscalar(obj1, obj2)
}
\arguments{
\item{obj1}{The existing \code{ssbrain} object}

\item{obj2}{The new \code{ssdscalar} object to add}
}
\description{
The internal function that calculates the result for \code{\link[ssbrain]{ss_dscalar}}
}
