% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_views.R
\name{add_view}
\alias{add_view}
\title{(Internal) Set a Viewpoint}
\usage{
add_view(obj1, obj2)
}
\arguments{
\item{obj1}{The existing \code{ssbrain} object}

\item{obj2}{The new \code{ssview} object to add}
}
\description{
The internal function that calculates the result for \code{\link[ssbrain]{ss_view}}
}
