% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_show_capture.R
\name{captureBrain}
\alias{captureBrain}
\title{Capture a Visually-displayed Brain Surface}
\usage{
captureBrain(
  brain,
  hemisphere,
  filename,
  width = 800,
  height = 500,
  crop = TRUE,
  cropmargin = 10
)
}
\arguments{
\item{brain}{An \code{ssbrain} object created by \link{ss_surf} (and potentially modified by other functions)}

\item{hemisphere}{Which brain hemisphere to display (either "left" or "right")}

\item{filename}{A string with the full path to the PNG file you'd like to create}

\item{width}{The width of the window. Default is 800.}

\item{height}{The height of the window. Default is 500.}

\item{crop}{Whether or not to crop the resulting image. Default is TRUE.}

\item{cropmargin}{The margin of whitespace to leave around the brain when cropping. Default is 10.}
}
\description{
This function opens an rgl window and displays a brain surface, then saves a screenshot of it.
}
\examples{
\dontrun{
my_brain = ss_surf(surf="fsaverage6") +
  ss_dscalar(dscalar_filename = "/path/to/my_file.dscalar.nii")

captureBrain(my_brain, hemisphere = "left", filename = "/path/to/my/output_image.png")
}
}
