% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_dscalar.R
\name{checkBin}
\alias{checkBin}
\title{Check data bins}
\usage{
checkBin(x, y)
}
\arguments{
\item{x}{The levels (bins) of the output of the \code{cut} function}

\item{y}{The value to check}
}
\value{
TRUE or FALSE
}
\description{
Check to see which bin contains the value
}
