% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_show_capture.R
\name{closeBrainViewers}
\alias{closeBrainViewers}
\title{Close All Open Brain Windows}
\usage{
closeBrainViewers()
}
\description{
Closes all currently-open brain windows
}
\examples{
\dontrun{
my_brain = ss_surf(surf="fsaverage6") +
  ss_dscalar(dscalar_filename = "/path/to/my_file.dscalar.nii")

showBrain(my_brain, hemisphere = "left")

closeBrainViewers()
}
}
