% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_colorbars.R
\name{colorbarGenerator}
\alias{colorbarGenerator}
\title{Generate Color Bar}
\usage{
colorbarGenerator(colorbar_name, neg_palette, pos_palette, full_palette)
}
\arguments{
\item{colorbar_name}{A string with the name of the color bar}

\item{neg_palette}{(Optional argument) A custom color palette created by colorRampPalette, to be used for negative values}

\item{pos_palette}{(Optional argument) A custom color palette created by colorRampPalette, to be used for positive values}
}
\description{
Given a name or color palette(s), generate the necessary information to use them in brain maps
}
