% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_colorbars.R
\name{createOutputList}
\alias{createOutputList}
\title{Create Color Maps}
\usage{
createOutputList(colorbar_name, pos_list, neg_list)
}
\arguments{
\item{colorbar_name}{A string with the name of the color bar}

\item{pos_list}{The list of positive color values to create the palette}

\item{neg_list}{The list of negative color values to create the palette}
}
\description{
Creates the necessary outputs for colormaps used by ss_dscalar and ss_seed
}
