% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_import_export.R
\name{distill_dconn}
\alias{distill_dconn}
\title{'Distill' a Connectivity Dconn File}
\usage{
distill_dconn(dconn_filepath, seed_value, num_verts)
}
\arguments{
\item{dconn_filepath}{A string with the full path to the .dconn.nii file}

\item{seed_value}{A value for the seed (the row number, not vertex number)}

\item{num_verts}{The number of vertices (row/columns) in your dconn file}
}
\value{
An object of the same type returned by \link{importCifti} when used on a dscalar file.
}
\description{
This function "distills" a symmetrical .dconn.nii file containing connectivity information. It exports the correlation between each vertex and the specified vertex, resulting in a vector the length of the number of vertices. Functionally, it converts the dconn into a dscalar.
}
\examples{
\dontrun{
dconn_seed_data = distill_dconn(dconn_filepath = "/path/to/my/dconn/my_data.dconn.nii",
                                seed_value = 2214, num_verts = 81924)
}
}
