% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_import_export.R
\name{exportCifti}
\alias{exportCifti}
\title{Export a CIFTI Data File or Border File}
\usage{
exportCifti(cifti_filename, data, luttpath, surf_l, surf_r, template)
}
\arguments{
\item{cifti_filename}{A string with the full path to the CIFTI data file you'd like to create}

\item{data}{A vector of data values.}

\item{luttpath}{(Optional argument) A path to a lutt color file; used for dlabel and border files. Default is the Yeo 17-parcel color scheme.}
}
\description{
This function exports a CIFTI data file (or border file) from R, using a vector of data
}
\details{
This function currently only supports \code{.dscalar.nii}, \code{.dlabel.nii}, and \code{.border} files.
For border files, outputs should be named filename.border. These will automatically be changed into left/right hemisphere files called filename_lh.border and filename_rh.border
}
\examples{
\dontrun{
my_data = runif(81924, 1.0, 100.0) # a vector of 81,924 random values between 1 and 100
exportCifti("/path/to/my/output.dscalar.nii", my_data)

my_data = sample(1:15, 81924) # a vector of 81,924 random integers between 1 and 15
exportCifti("/path/to/my/output.dlabel.nii", my_data)
}

}
