% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_show_capture.R
\name{get_viewmatrix}
\alias{get_viewmatrix}
\title{Calculation of View Matrix}
\usage{
get_viewmatrix(hemisphere, side, wb_X, wb_Y, wb_Z)
}
\arguments{
\item{hemisphere}{"left" or "right"}

\item{side}{"lateral" or "medial"}

\item{wb_X}{The x-axis rotation as shown in Connectome Workbench}

\item{wb_Y}{The y-axis rotation as shown in Connectome Workbench}

\item{wb_Z}{The z-axis rotation as shown in Connectome Workbench}
}
\value{
A view matrix that can be passed to rgl::view3d as a userMatrix.
}
\description{
This function calculates the coordinates that \code{rgl} needs to rotate a brain.
}
