% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_seed_vector.R
\name{getSeedVector}
\alias{getSeedVector}
\title{Get a Linear Vector of Seeds}
\usage{
getSeedVector(brain, seed_start, seed_end, dlabel)
}
\arguments{
\item{brain}{Any \code{ss_surf} object in the same mesh space that you plan to use for \code{ss_seed} (i.e. fsaverage6, fsaverage7)}

\item{seed_start}{The value of the starting seed to draw from}

\item{seed_end}{The value of the ending seed to draw toward}

\item{dlabel}{(Optional argument)}
}
\value{
A list with three items: vertices (Euclidean coordinates for each vertex), faces (the 3 vertex indices that make up each face), and num_verts (the total number of vertices in the mesh)
}
\description{
This function calculates the values of every seed that lie in a (relatively) straight line between the two seeds provided. It allows you to "draw" a series of seeds across the cortex, by looping through the resulting list with \code{ss_seed}.
}
\examples{
\dontrun{
my_brain = ss_surf(surf = "fsaverage6")
my_seed_vector = getSeedVector(my_brain, 33424, 13264)
}
}
