% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_import_export.R
\name{importBorder}
\alias{importBorder}
\title{Import a Border File}
\usage{
importBorder(border_name, border_surfname)
}
\arguments{
\item{border_name}{A string with the full path to a border file}

\item{border_surfname}{A string with the full path to the GIFTI surface used to create the border}
}
\value{
A gifti file containing border information
}
\description{
This function imports a border file into R, providing information on vertices (and their ordering) used to form borders.
}
\examples{
\dontrun{
border_data = importBorder("/path/to/border/my_borderfile_lh.border",
                           border_surfname = "/path/to/surface/lh.pial_infl2.surf.gii")
}
}
