% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_import_export.R
\name{importCifti}
\alias{importCifti}
\title{Import a CIFTI Data File}
\usage{
importCifti(cifti_name, data_only = FALSE)
}
\arguments{
\item{cifti_name}{A string with the full path to a CIFTI data file}

\item{data_only}{If TRUE, only provides a vector of the data. If FALSE, provides a full \code{gifti} object, including metadata. Default is FALSE.}
}
\value{
A gifti object (data_only = FALSE) or a vector of data (data_only = TRUE)
}
\description{
This function imports a CIFTI data file into R, providing information for each surface vertex (depending on the filetype).
}
\details{
This function currently only supports \code{.dscalar.nii}, \code{.dlabel.nii}, and \code{.dconn.nii} files.
}
\examples{
\dontrun{
cifti = importCifti("/path/to/cifti/my_file.dlabel.nii")

# If `data_only` is FALSE (the default), the data inside the resulting
# object is located in `$data$normal`. For example:

cifti_object = importCifti("/path/to/cifti/my_cifti.dscalar.nii")
cifti_data = cifti_object$data$normal

# is equivalent to

cifti_data = importCifti("/path/to/cifti/my_cifti.dscalar.nii",
                          data_only = TRUE)
}

}
