% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_import_export.R
\name{importSurface}
\alias{importSurface}
\title{Import a GIFTI Surface File}
\usage{
importSurface(surface_name)
}
\arguments{
\item{surface_name}{A string with the full path to a GIFTI surface file}
}
\value{
A list with three items: vertices (Euclidean coordinates for each vertex), faces (the 3 vertex indices that make up each face), and num_verts (the total number of vertices in the mesh)
}
\description{
This function imports a GIFTI surface file into R, providing information for vertex coordinates in Euclidean space and the vertices used to draw triangular faces.
}
\examples{
\dontrun{
surface_data = importSurface("/path/to/surface/lh.pial_infl2.surf.gii")
}
}
