% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_border.R
\name{is.border}
\alias{is.border}
\title{Check if Border}
\usage{
is.border(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
TRUE or FALSE
}
\description{
This function checks if an object is of the class \code{ssborder}.
}
\examples{
\dontrun{
new_border = ss_border(filename = "/path/to/my/datafile.dlabel.nii", hemisphere = "left")
is.border(new_border)
}
}
