% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_seed.R
\name{is.dconn}
\alias{is.dconn}
\title{Check if Dconn}
\usage{
is.dconn(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
TRUE or FALSE
}
\description{
This function checks if an object is of the class \code{ssdconn}.
}
\examples{
\dontrun{
new_dconn = ss_seed(dconn_filename = "/path/to/my/datafile.dconn.nii", seed_value = 2213)
is.dconn(new_dconn)
}
}
