% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_dlabel.R
\name{is.dlabel}
\alias{is.dlabel}
\title{Check if Dlabel}
\usage{
is.dlabel(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
TRUE or FALSE
}
\description{
This function checks if an object is of the class \code{ssdlabel}.
}
\examples{
\dontrun{
new_dlabel = ss_dlabel(filename = "/path/to/my/datafile.dlabel.nii")
is.dlabel(new_dlabel)
}
}
