% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_dscalar.R
\name{is.dscalar}
\alias{is.dscalar}
\title{Check if Dscalar}
\usage{
is.dscalar(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
TRUE or FALSE
}
\description{
This function checks if an object is of the class \code{ssdscalar}.
}
\examples{
\dontrun{
new_dscalar = ss_dscalar(dscalar_filename = "/path/to/my/datafile.dscalar.nii")
is.dscalar(new_dscalar)
}
}
