% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_views.R
\name{is.view}
\alias{is.view}
\title{Check if View}
\usage{
is.view(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
TRUE or FALSE
}
\description{
This function checks if an object is of the class \code{ssview}.
}
\examples{
\dontrun{
new_view = ss_view(side="lateral", rotation = "inferior_temporal")
is.view(new_view)
}
}
