% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_colorbars.R
\name{listColorbars}
\alias{listColorbars}
\title{List Available Colorbars}
\usage{
listColorbars()
}
\description{
This function lists the colorbars available in the \code{ssbrain} package for use with the \code{\link[ssbrain]{ss_dscalar}} and \code{\link[ssbrain]{ss_seed}} functions.
}
