% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_views.R
\name{listRotations}
\alias{listRotations}
\title{List Available Rotations}
\usage{
listRotations()
}
\description{
This function lists the pre-set rotations available in the \code{ssbrain} package for use with the \code{\link[ssbrain]{ss_view}} function.
}
