% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_plus_function.R
\name{+.ssbrain}
\alias{+.ssbrain}
\title{Overloaded Plus}
\usage{
\method{+}{ssbrain}(obj1, obj2)
}
\arguments{
\item{obj1}{The existing \code{ssbrain} object}

\item{obj2}{The new \code{ssbrain} object to add}
}
\value{
A new \code{ssbrain} object with information from both previous objects.
}
\description{
This is an overloaded function for '+' that allows for combining different \code{ssbrain} elements.
}
