% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_colorbars.R
\name{rgb_fun1}
\alias{rgb_fun1}
\title{Create a Color Code}
\usage{
rgb_fun1(x)
}
\arguments{
\item{x}{A vector with three RGB values}
}
\description{
Create a color HEX code from an RGB triple in a vector
}
