% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_wbpath.R
\name{set_wbpath}
\alias{set_wbpath}
\title{Set Workbench Directory}
\usage{
set_wbpath(value)
}
\arguments{
\item{value}{A string with the full filepath to your Workbench executable. It should end with "wb_command"}
}
\description{
This allows you to set the path to your Workbench executable file.
}
\details{
This sets a global option for your R environment. It only needs to be run the first time you start R.
It is advised that you put this function at the top of every ssbrain script.
}
\examples{
\dontrun{
set_wbpath("/path/to/directory/for/workbench/wb_command")
}

}
