% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_show_capture.R
\name{showBrain}
\alias{showBrain}
\title{Visually Display Brain Surface}
\usage{
showBrain(brain, hemisphere, width = 800, height = 500)
}
\arguments{
\item{brain}{An \code{ssbrain} object created by \link{ss_surf} (and potentially modified by other functions)}

\item{hemisphere}{Which brain hemisphere to display (either "left" or "right")}

\item{width}{The width of the window. Default is 800.}

\item{height}{The height of the window. Default is 500.}
}
\description{
This function opens an rgl window and displays a brain surface.
}
\examples{
\dontrun{
my_brain = ss_surf(surf="fsaverage6") +
  ss_dscalar(dscalar_filename = "/path/to/my_file.dscalar.nii")

showBrain(my_brain, hemisphere = "left")
}
}
