% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_dscalar.R
\name{ss_dscalar}
\alias{ss_dscalar}
\title{Set a Dscalar}
\usage{
ss_dscalar(
  dscalar_filename,
  colorbar = "FSL",
  show = "all",
  pos_thresh = 1,
  neg_thresh = -1,
  pos_colorrange = c(1, 3),
  neg_colorrange = c(-1, -3),
  pos_palette,
  neg_palette
)
}
\arguments{
\item{dscalar_filename}{A string with the full filepath to a dscalar file}

\item{colorbar}{A string containing the name of a colorbar to use (same names as Connectome Workbench)}

\item{show}{Whether to show positive data, negative data, or all data. Options are "all", "pos", or "neg". Defaults to "all".}

\item{pos_thresh}{The threshold below which to hide positive data. Defaults to 1.}

\item{neg_thresh}{The option above which to hide negative data. Defaults to -1.}

\item{pos_colorrange}{The positive range in which to spread the \code{colorbar}'s colors. Defaults to c(1,3)}

\item{neg_colorrange}{The negative range in which to spread the \code{colorbar}'s colors. Defaults to c(-1,-3)}

\item{pos_palette}{(Experimental) A color palette created by \code{colorRampPalette}; overrides \code{colorbar} for positive colors.}

\item{neg_palette}{(Experimental) A color palette created by \code{colorRampPalette}; overrides \code{colorbar} for negative colors.}
}
\description{
This function sets a dscalar data file on the surface of the brain
}
\details{
Note: Adding multiple \code{ss_dscalar} items to the same object will cause overlapping brain maps, with the most recently-added on top.
}
\examples{
\dontrun{
my_brain = ss_surf(surf="fsaverage6") +
  ss_dscalar(dscalar_filename = "/path/to/my/datafile.dscalar.nii")

my_brain = ss_surf(surf="fsaverage6")

my_brain_new = my_brain +
  ss_dscalar(dscalar_filename = "/path/to/my/datafile.dscalar.nii",
             colorbar = "ROY-BIG-BL",
             show= "pos",
             pos_thresh = 10,
             pos_colorrange = c(10,50))
}

}
