% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbrain_surf.R
\name{ss_surf}
\alias{ss_surf}
\title{Set a Surface Mesh}
\usage{
ss_surf(
  surf = NULL,
  surfL = NULL,
  surfR = NULL,
  showMedialWall = TRUE,
  medialWallColor = "#858585"
)
}
\arguments{
\item{surf}{A string with the surface mesh you wish to use. Options are "fsaverage6" and "fsaverage7".}

\item{surfL}{(Optional argument) If you don't want to use fs6/fs7, you can provide the full path to a surface GIFTI file for the left hemisphere.}

\item{surfR}{(Optional argument) If you don't want to use fs6/fs7, you can provide the full path to a surface GIFTI file for the right hemisphere.}
}
\value{
An \code{ssbrain} object with the specified surface mesh(es).
}
\description{
This is the core function of the \code{ssbrain} package. It creates an object that will store all of the surface information you want to display. It sets which surface mesh you plan to use.
}
\examples{
\dontrun{
my_brain = ss_surf(surf="fsaverage6")

my_brain = ss_surf(surfL = "/path/to/surface/lh.pial_infl2.surf.gii")
}
}
